code/requirements.txt
pandas
requests
gitpython
gql
python-dotenv
